Use inference from cropharvest
